Added errors and submitErrors to FormState
