+++ src/formSubscriptionItems.js
@@ -3,11 +3,13 @@ export default [
   'active',
   'dirty',
   'error',
+  'errors',
   'initialValues',
   'invalid',
   'pristine',
   'submitting',
   'submitError',
+  'submitErrors',
   'submitFailed',
   'submitSucceeded',
   'valid',
