+++ src/types.js.flow
@@ -6,10 +6,12 @@ export type FormSubscription = $Shape<{
   active: boolean,
   dirty: boolean,
   error: boolean,
+  errors: boolean,
   initialValues: boolean,
   invalid: boolean,
   pristine: boolean,
   submitting: boolean,
+  submitError: boolean,
   submitFailed: boolean,
   submitSucceeded: boolean,
   valid: boolean,
@@ -23,6 +25,7 @@ export type FormState = $Shape<{
   active: string,
   dirty: boolean,
   error: any,
+  errors: Object,
   invalid: boolean,
   initialValues: Object,
   pristine: boolean,
@@ -30,6 +33,7 @@ export type FormState = $Shape<{
   submitFailed: boolean,
   submitSucceeded: boolean,
   submitError: any,
+  submitErrors: Object,
   valid: boolean,
   validating: boolean,
   values: Object
